+++ src/apps/umami-finance/arbitrum/umami-finance.glp-vaults.token-fetcher.ts
@@ -2,6 +2,7 @@ import { Inject } from '@nestjs/common';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+import { getTokenImg } from '~app-toolkit/helpers/presentation/image.present';
 import { AppTokenTemplatePositionFetcher } from '~position/template/app-token.template.position-fetcher';
 import {
   GetAddressesParams,
@@ -10,15 +11,18 @@ import {
   GetPricePerShareParams,
   DefaultAppTokenDataProps,
   GetDisplayPropsParams,
+  GetDataPropsParams,
 } from '~position/template/app-token.template.types';
 
 import { UmamiFinanceGlpVaultAddress } from '../common/umami-finance.constants';
+import { UmamiFinanceYieldResolver } from '../common/umami-finance.yield-resolver';
 import { UmamiFinanceContractFactory } from '../contracts';
 import { UmamiFinanceGlpVault } from '../contracts/ethers/UmamiFinanceGlpVault';
 
 export type UmamiFinanceGlpVaultAppTokenDefinition = {
   address: string;
   timelockedVaultAddress: string;
+  apiId: string;
 };
 
 @PositionTemplate()
@@ -27,14 +31,15 @@ export class ArbitrumUmamiFinanceGlpVaultsTokenFetcher extends AppTokenTemplateP
   DefaultAppTokenDataProps,
   UmamiFinanceGlpVaultAppTokenDefinition
 > {
-  groupLabel: string;
+  groupLabel = 'GLP Vaults';
 
   constructor(
     @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
+    @Inject(UmamiFinanceYieldResolver)
+    private readonly yieldResolver: UmamiFinanceYieldResolver,
     @Inject(UmamiFinanceContractFactory) private readonly umamiFinanceContractFactory: UmamiFinanceContractFactory,
   ) {
     super(appToolkit);
-    this.groupLabel = 'GLP Vaults';
   }
 
   getContract(_address: string): UmamiFinanceGlpVault {
@@ -76,5 +81,21 @@ export class ArbitrumUmamiFinanceGlpVaultsTokenFetcher extends AppTokenTemplateP
 
   async getLabel({ appToken }: GetDisplayPropsParams<UmamiFinanceGlpVault>): Promise<string> {
     return `GLP ${appToken.tokens[0].symbol}`;
+  }
+
+  async getImages({
+    appToken,
+  }: GetDisplayPropsParams<
+    UmamiFinanceGlpVault,
+    DefaultAppTokenDataProps,
+    UmamiFinanceGlpVaultAppTokenDefinition
+  >): Promise<string[]> {
+    return [getTokenImg(appToken.address, this.network)];
+  }
+
+  async getApy(
+    _params: GetDataPropsParams<UmamiFinanceGlpVault, DefaultAppTokenDataProps, UmamiFinanceGlpVaultAppTokenDefinition>,
+  ): Promise<number> {
+    return this.yieldResolver.getVaultYield(_params.address);
   }
 }