+++ src/apps/vendor-finance/common/vendor-finance.pool.contract-position-fetcher.ts
@@ -1,4 +1,5 @@
 import { Inject } from '@nestjs/common';
+import Axios from 'axios';
 
 import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import {
@@ -6,7 +7,6 @@ import {
   buildPercentageDisplayItem,
 } from '~app-toolkit/helpers/presentation/display-item.present';
 import { getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
-import { gqlFetch } from '~app-toolkit/helpers/the-graph.helper';
 import { DefaultDataProps } from '~position/display.interface';
 import { MetaType } from '~position/position.interface';
 import { isBorrowed } from '~position/position.utils';
@@ -20,17 +20,15 @@ import {
 
 import { VendorFinanceContractFactory, VendorFinancePool } from '../contracts';
 
-import { borrowerInfosQuery } from './getBorrowerInfosQuery';
 import { LENDING_POOLS_QUERY } from './getLendingPoolsQuery';
 import {
-  VendorBorrowerGraphResponse,
   VendorFinancePoolDataProps,
   VendorFinancePoolDefinition,
   VendorLendingPoolsGraphResponse,
 } from './vendor-finance.pool.types';
 
 export abstract class VendorFinancePoolContractPositionFetcher extends ContractPositionTemplatePositionFetcher<VendorFinancePool> {
-  abstract subgraphUrl: string;
+  abstract entityUrl: string;
 
   constructor(
     @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
@@ -44,12 +42,11 @@ export abstract class VendorFinancePoolContractPositionFetcher extends ContractP
   }
 
   async getDefinitions() {
-    const data = await gqlFetch<VendorLendingPoolsGraphResponse>({
-      endpoint: this.subgraphUrl,
+    const data: VendorLendingPoolsGraphResponse = await Axios.post(this.entityUrl, {
       query: LENDING_POOLS_QUERY,
+      type: `v1-${this.network.charAt(0).toUpperCase() + this.network.slice(1)}`,
     });
-
-    return (data?.pools ?? []).map(poolData => ({
+    return (data.data?.pools ?? []).map(poolData => ({
       address: poolData.id,
       deployer: poolData._deployer,
       mintRatio: poolData._mintRatio,
@@ -122,17 +119,22 @@ export abstract class VendorFinancePoolContractPositionFetcher extends ContractP
       },
     ];
   }
-  async getDataProps(
-    _params: GetDataPropsParams<VendorFinancePool, DefaultDataProps, VendorFinancePoolDefinition>,
-  ): Promise<VendorFinancePoolDataProps> {
+  async getDataProps({
+    definition,
+  }: GetDataPropsParams<
+    VendorFinancePool,
+    DefaultDataProps,
+    VendorFinancePoolDefinition
+  >): Promise<VendorFinancePoolDataProps> {
     return {
-      deployer: _params.definition.deployer,
-      totalDeposited: parseInt(_params.definition.lendBalance) + parseInt(_params.definition.totalBorrowed),
+      deployer: definition.deployer,
+      totalDeposited: parseInt(definition.lendBalance) + parseInt(definition.totalBorrowed),
     };
   }
 
   async getTokenBalancesPerPosition({
     address,
+    contract,
     contractPosition,
   }: GetTokenBalancesParams<VendorFinancePool, VendorFinancePoolDataProps>) {
     const collateralToken = contractPosition.tokens[0]!;
@@ -146,18 +148,12 @@ export abstract class VendorFinancePoolContractPositionFetcher extends ContractP
     // --! Lender logic !---
 
     // --- Borrower logic ----
-    const data = await gqlFetch<VendorBorrowerGraphResponse>({
-      endpoint: this.subgraphUrl,
-      query: borrowerInfosQuery(address),
-    });
+    const [borrowPosition, mintRatioRaw] = await Promise.all([contract.debt(address), contract.mintRatio()]);
 
-    const borrowerPosition = data.borrower?.positions.find(({ pool }) => pool.id === contractPosition.address);
-    if (!borrowerPosition) return [];
-
-    const poolRate = parseInt(borrowerPosition.pool._mintRatio) / 10 ** 18;
-    const suppliedBalance = parseInt(borrowerPosition.totalBorrowed) / 10 ** lentToken.decimals / poolRate;
+    const poolRate = Number(mintRatioRaw) / 10 ** 18;
+    const suppliedBalance = Number(borrowPosition.borrowAmount) / 10 ** lentToken.decimals / poolRate;
     const suppliedBalanceRaw = suppliedBalance * 10 ** collateralToken.decimals;
-    const borrowedBalance = parseInt(borrowerPosition.totalBorrowed);
+    const borrowedBalance = Number(borrowPosition.borrowAmount);
 
     // Deposit, borrow, no lending out (not pool creator)
     return [suppliedBalanceRaw.toString(), borrowedBalance.toString(), '0'];