+++ src/apps/vendor-finance/common/vendor-finance.pool.types.ts
@@ -33,25 +33,8 @@ export type VendorLendingPool = {
 };
 
 export type VendorLendingPoolsGraphResponse = {
+  data: any;
   pools: Array<VendorLendingPool>;
-};
-
-// Used for V1 Pools
-export type BorrowerPosition = {
-  pool: {
-    id: string;
-    _mintRatio: string;
-  };
-  totalBorrowed: string;
-  effectiveRate: string;
-};
-
-export type VendorBorrower = {
-  positions: Array<BorrowerPosition>;
-};
-
-export type VendorBorrowerGraphResponse = {
-  borrower: VendorBorrower;
 };
 
 export type VendorFinancePoolV2Definition = {
@@ -82,5 +65,6 @@ export type VendorLendingPoolV2 = {
 };
 
 export type VendorLendingPoolsV2GraphResponse = {
+  [x: string]: any;
   pools: Array<VendorLendingPoolV2>;
 };