+++ src/apps/yearn/contracts/ethers/index.ts
@@ -8,6 +8,7 @@ export type { YearnStaking } from './YearnStaking';
 export type { YearnStakingRewardRegistry } from './YearnStakingRewardRegistry';
 export type { YearnVault } from './YearnVault';
 export type { YearnVaultV2 } from './YearnVaultV2';
+export type { YearnVeYfi } from './YearnVeYfi';
 export type { YearnYCrv } from './YearnYCrv';
 export * as factories from './factories';
 export { YearnGovernance__factory } from './factories/YearnGovernance__factory';
@@ -17,4 +18,5 @@ export { YearnStakingRewardRegistry__factory } from './factories/YearnStakingRew
 export { YearnStaking__factory } from './factories/YearnStaking__factory';
 export { YearnVaultV2__factory } from './factories/YearnVaultV2__factory';
 export { YearnVault__factory } from './factories/YearnVault__factory';
+export { YearnVeYfi__factory } from './factories/YearnVeYfi__factory';
 export { YearnYCrv__factory } from './factories/YearnYCrv__factory';