+++ src/types/network.interface.ts
@@ -14,6 +14,7 @@ export enum Network {
   CRONOS_MAINNET = 'cronos',
   AURORA_MAINNET = 'aurora',
   EVMOS_MAINNET = 'evmos',
+  BASE_MAINNET = 'base',
 }
 
 export const NETWORK_IDS: Record<Network, number | null> = {
@@ -32,4 +33,5 @@ export const NETWORK_IDS: Record<Network, number | null> = {
   [Network.AURORA_MAINNET]: 1313161554,
   [Network.BITCOIN_MAINNET]: 1285,
   [Network.EVMOS_MAINNET]: 9001,
+  [Network.BASE_MAINNET]: 8453,
 };